learning-platform/requirements.txt
Flask==1.1.2
google-cloud-datastore==1.12.0
google-cloud-translate==2.0.1
google-cloud-storage==1.28.1
google-auth==1.15.0
requests==2.23.0